+++ lib/td_dq/canada/abilities.ex
@@ -1,8 +1,11 @@
 defmodule TdDq.Canada.Abilities do
   @moduledoc false
+  alias TdCache.ConceptCache
   alias TdDq.Accounts.User
   alias TdDq.Canada.RuleAbilities
   alias TdDq.Canada.RuleImplementationAbilities
+  alias TdDq.Permissions
+  alias TdDq.Rules.Rule
   alias TdDq.Rules.RuleImplementation
 
   defimpl Canada.Can, for: User do
@@ -20,46 +23,85 @@ defmodule TdDq.Canada.Abilities do
       RuleImplementationAbilities.can?(user, action, rule_implementation)
     end
 
+    def can?(%User{} = user, :manage, %{
+          "business_concept_id" => nil,
+          "resource_type" => "rule_implementation"
+        }) do
+      RuleImplementationAbilities.can?(user, :manage_quality_rule_implementations, "")
+    end
+
+    def can?(%User{} = user, :manage, %{
+          "business_concept_id" => business_concept_id,
+          "resource_type" => "rule_implementation"
+        }) do
+      RuleImplementationAbilities.can?(
+        user,
+        :manage_quality_rule_implementations,
+        business_concept_id
+      )
+    end
+
+    def can?(%User{} = user, :show, %Rule{business_concept_id: nil}) do
+      RuleAbilities.can?(user, :show, "")
+    end
+
+    def can?(%User{} = user, :show, %Rule{business_concept_id: business_concept_id}) do
+      RuleAbilities.can?(user, :show, business_concept_id) &&
+        authorized?(user, business_concept_id)
+    end
+
     def can?(%User{} = user, :create, %{
           "business_concept_id" => nil,
           "resource_type" => "rule_implementation"
         }) do
-      RuleImplementationAbilities.can?(user, :manage_rules, "")
+      RuleImplementationAbilities.can?(user, :manage_quality_rule_implementations, "")
     end
 
     def can?(%User{} = user, :create, %{
           "business_concept_id" => business_concept_id,
           "resource_type" => "rule_implementation"
         }) do
-      RuleImplementationAbilities.can?(user, :manage_rules, business_concept_id)
+      RuleImplementationAbilities.can?(
+        user,
+        :manage_quality_rule_implementations,
+        business_concept_id
+      )
     end
 
     def can?(%User{} = user, :update, %{
           "business_concept_id" => nil,
           "resource_type" => "rule_implementation"
         }) do
-      RuleImplementationAbilities.can?(user, :manage_rules, "")
+      RuleImplementationAbilities.can?(user, :manage_quality_rule_implementations, "")
     end
 
     def can?(%User{} = user, :update, %{
           "business_concept_id" => business_concept_id,
           "resource_type" => "rule_implementation"
         }) do
-      RuleImplementationAbilities.can?(user, :manage_rules, business_concept_id)
+      RuleImplementationAbilities.can?(
+        user,
+        :manage_quality_rule_implementations,
+        business_concept_id
+      )
     end
 
     def can?(%User{} = user, :delete, %{
           "business_concept_id" => nil,
           "resource_type" => "rule_implementation"
         }) do
-      RuleImplementationAbilities.can?(user, :manage_rules, "")
+      RuleImplementationAbilities.can?(user, :manage_quality_rule_implementations, "")
     end
 
     def can?(%User{} = user, :delete, %{
           "business_concept_id" => business_concept_id,
           "resource_type" => "rule_implementation"
         }) do
-      RuleImplementationAbilities.can?(user, :manage_rules, business_concept_id)
+      RuleImplementationAbilities.can?(
+        user,
+        :manage_quality_rule_implementations,
+        business_concept_id
+      )
     end
 
     def can?(%User{} = user, :manage, %{"resource_type" => "rule"}) do
@@ -70,7 +112,8 @@ defmodule TdDq.Canada.Abilities do
           "business_concept_id" => business_concept_id,
           "resource_type" => "rule"
         }) do
-      RuleAbilities.can?(user, :index_rule, business_concept_id)
+      RuleAbilities.can?(user, :index_rule, business_concept_id) &&
+        authorized?(user, business_concept_id)
     end
 
     def can?(%User{} = user, :create, %{
@@ -84,7 +127,8 @@ defmodule TdDq.Canada.Abilities do
           "business_concept_id" => business_concept_id,
           "resource_type" => "rule"
         }) do
-      RuleAbilities.can?(user, :manage_rules, business_concept_id)
+      RuleAbilities.can?(user, :manage_rules, business_concept_id) &&
+        authorized?(user, business_concept_id)
     end
 
     def can?(%User{} = user, :update, %{
@@ -98,7 +142,8 @@ defmodule TdDq.Canada.Abilities do
           "business_concept_id" => business_concept_id,
           "resource_type" => "rule"
         }) do
-      RuleAbilities.can?(user, :manage_rules, business_concept_id)
+      RuleAbilities.can?(user, :manage_rules, business_concept_id) &&
+        authorized?(user, business_concept_id)
     end
 
     def can?(%User{} = user, :delete, %{
@@ -112,9 +157,26 @@ defmodule TdDq.Canada.Abilities do
           "business_concept_id" => business_concept_id,
           "resource_type" => "rule"
         }) do
-      RuleAbilities.can?(user, :manage_rules, business_concept_id)
+      RuleAbilities.can?(user, :manage_rules, business_concept_id) &&
+        authorized?(user, business_concept_id)
     end
 
     def can?(%User{}, _action, _entity), do: false
+
+    defp authorized?(%User{} = user, business_concept_id) do
+      {:ok, status} = ConceptCache.member_confidential_ids(business_concept_id)
+
+      case status do
+        1 ->
+          Permissions.authorized?(
+            user,
+            :manage_confidential_business_concepts,
+            business_concept_id
+          )
+
+        _ ->
+          true
+      end
+    end
   end
 end