+++ lib/td_dq/permissions/permissions.ex
@@ -26,4 +26,8 @@ defmodule TdDq.Permissions do
     @permission_resolver.has_permission?(jti, permission)
   end
 
+  def get_domain_permissions(%User{jti: jti}) do
+    @permission_resolver.get_acls_by_resource_type(jti, "domain")
+  end
+
 end