+++ test/features/rule_test.exs
@@ -6,10 +6,12 @@ defmodule TdQd.RuleTest do
   import TdDqWeb.Rule, only: :functions
   import TdDqWeb.RuleType, only: :functions
   import TdDqWeb.ResponseCode, only: :functions
+  alias TdDq.Permissions.MockPermissionResolver
   alias TdDqWeb.ApiServices.MockTdAuditService
 
   setup_all do
     start_supervised(MockTdAuditService)
+    start_supervised(MockPermissionResolver)
     :ok
   end